silent doautocmd
The exists() check is unreliable if advanced pattern constructs are used
to define an autocommand. For example:

  autocommand User Foo{Bar,Baz} call handler()

Running exists('#User#FooBar') will return false. This is not a huge
issue since this plugin only broadcasts a single pattern, but the other
aspect of the issue is that it's acceptable to use "silent" to broadcast
the event unconditionally, since most handlers aren't going to be
displaying messages, and if they absolutely have to, they can use
"unsilent".